generate.js разделен на 2 файла
